feat: extraction of exporter functions from SimplePanel.tsx to separate files and proper typing for them
