+++ src/components/SimplePanel.tsx
@@ -1,13 +1,8 @@
 import { css, cx } from '@emotion/css';
 import { GrafanaTheme2, PanelProps, StandardEditorProps } from '@grafana/data';
-import { DateTime } from "luxon";
 import React, { useEffect, useRef, useState } from 'react';
 import ReactJson from 'react-json-view';
 import './../../ngx-flow/widget/ngx-flow.js';
-import { Buffer } from 'buffer';
-// @ts-ignore
-import { configure } from 'pcap-generator';
-import { encodeEthernetFrame } from 'helpers/packets/ethernetPacket';
 
 import {
     Button,
@@ -19,14 +14,10 @@ import {
     useStyles2,
     useTheme2
 } from '@grafana/ui';
-import { convertDateToFileName } from 'helpers/convertDateToFileName';
-import { hash } from 'helpers/hash';
-import { labelFormatter } from 'helpers/labelFormatter';
+import { filterFlowItems } from 'helpers/dataProcessors/filterFlowItems';
+import { pcapExporter, textExporter } from 'helpers/exporters';
 import { CopyText } from './CopyText/CopyText';
 import { FilterPanel, Filters } from './FilterPanel/FilterPanel';
-import { encodeIPV4Packet, encodeIPV6Packet } from 'helpers/packets/ipPacket';
-import { encodeUDPFrame } from 'helpers/packets/udpPacket';
-import { encodeTCPFrame } from 'helpers/packets/tcpPacket';
 
 
 
@@ -210,50 +201,6 @@ let ngxFlowClickHandler: Function = function () { };
 document.addEventListener('ngx-flow-click-item', function (e: any) {
     ngxFlowClickHandler(e)
 });
-
-function formattingDataAndSortIt(data: any, sortType = 'none') {
-
-    let [firstField] = data || [];
-
-    const unSortData = firstField?.values?.map((i: any, k: number) => {
-        const outData: any = {};
-        data.forEach((item: any) => {
-            outData[item.name] = item?.values?.[k];
-        });
-        if (outData?.Time && typeof outData?.labels === 'object') {
-            outData['labels'].timestamp = outData.Time;
-        }
-        return outData;
-    }) || [];
-    if (sortType === 'none') {
-        return unSortData;
-    }
-    const sortData = unSortData.sort((itemA: any, itemB: any) => {
-        if (itemA.tsNs && itemB.tsNs) {
-            const a = itemA.tsNs;
-            const b = itemB.tsNs;
-            return a < b ? -1 : a > b ? 1 : 0;
-        } else {
-            const a = itemA.Time;
-            const b = itemB.Time;
-            return a < b ? -1 : a > b ? 1 : 0;
-        }
-    });
-    if (sortType === 'time_old') {
-        return sortData;
-    }
-
-    if (sortType === 'time_new') {
-        return sortData.reverse();
-    }
-
-}
-
-interface Labels {
-    [key: string]: string
-}
-
-
 export const SimplePanel: React.FC<Props> = ({ options, data, width, height }) => {
     const [flowData, setFlowData] = React.useState({ actors: [], data: [] });
     const [modalIsOpen, setModalIsOpen] = React.useState(false);
@@ -266,37 +213,7 @@ export const SimplePanel: React.FC<Props> = ({ options, data, width, height }) =
     // Export .txt
     useEffect(() => {
         const handler = (e: any) => {
-
-            const [serie]: any = (data as any)?.series || [];
-            const fields = serie?.fields || [];
-            const lineField = fields.find((i: any) => i.name === 'Line') ?? [];
-            const exportText = lineField?.values.map((i: string, index: number) => {
-                let dt = DateTime.fromMillis(fields[0]?.values[index]?.timestamp).toISO()
-                dt = dt?.replace('+', '000+') ?? '';
-                if (fields[0]?.values[index]?.type !== 'sip') {
-                    i = '';
-                }
-                let proto = 'proto:'
-                if (i.includes('UDP')) {
-                    proto += "UDP"
-                } else if (i.includes('TCP')) {
-                    proto += "TCP"
-                }
-                return `${proto} ${dt} ${fields[0]?.values[index]?.src_ip} ---> ${fields[0]?.values[index]?.dst_ip} \n\n${i}`
-            }).join('\n');
-            // Create element with <a> tag
-            const link = document.createElement("a");
-
-            // Create a blog object with the file content which you want to add to the file
-            const file = new Blob([exportText], { type: 'text/plain' });
-
-            // Add file content in the object URL
-            link.href = URL.createObjectURL(file);
-
-            // Add file name
-            const date = new Date();
-            link.download = `${convertDateToFileName(date)}.txt`;
-            link.click();
+            textExporter(data);
         }
         document.addEventListener('export-flow-as-text', handler);
         return () => {
@@ -306,86 +223,7 @@ export const SimplePanel: React.FC<Props> = ({ options, data, width, height }) =
     // Export .pcap
     useEffect(() => {
         const handler = (e: any) => {
-
-            const [serie]: any = (data as any)?.series || [];
-            const fields = serie?.fields || [];
-            const lineField = fields.find((i: any) => i.name === 'Line') ?? [];
-            const values = fields[0]?.values.map((item: any, index: number) => {
-                item.line = lineField?.values[index];
-                return item
-            }).filter((packet: any) => packet.type === 'sip')
-            const sequenceMap = new Map<string, number>();
-            const ipv4_regex = /(?:25[0-5]|2[0-4]\d|1\d\d|[1-9]\d|\d)(?:\.(?:25[0-5]|2[0-4]\d|1\d\d|[1-9]\d|\d)){3}/;
-            const packets2 = values.sort((a: any, b: any) => {
-                return a.timestamp - b.timestamp
-            }).map((labels: Labels, index: number) => {
-                let line = labels.line
-                let proto = ''
-                if (line.includes('UDP')) {
-                    proto += "UDP"
-                } else if (line.includes('TCP')) {
-                    proto += "TCP"
-                }
-                const fieldObj = {
-                    data: line,
-                    srcIp: labels.src_ip,
-                    dstIp: labels.dst_ip,
-                    srcPort: labels.src_port,
-                    dstPort: labels.dst_port,
-                    ts: fields[0]?.values[index]?.timestamp,
-                    proto: proto === 'UDP' ? 17 : 6 as 6 | 17,
-                    type: fields[0]?.values[index]?.type
-                }
-                const hash = `${fieldObj.srcIp}:${fieldObj.srcPort}->${fieldObj.dstIp}:${fieldObj.dstPort}`
-                const sequence = sequenceMap.get(hash) ?? 1
-                const packet_data = proto === 'UDP' ? encodeUDPFrame({
-                    sourcePort: fieldObj.srcPort,
-                    destinationPort: fieldObj.dstPort,
-                    data: Buffer.from(line)
-                }) : encodeTCPFrame({
-                    sourcePort: fieldObj.srcPort,
-                    destinationPort: fieldObj.dstPort,
-                    sequenceNumber: sequence,
-                    data: Buffer.from(line)
-                })
-                let ip_packet = ipv4_regex.test(fieldObj.dstIp) ? encodeIPV4Packet({
-                    protocol: fieldObj.proto,
-                    sourceIp: fieldObj.srcIp,
-                    destinationIp: fieldObj.dstIp,
-                    data: packet_data,
-                }) : encodeIPV6Packet({
-                    protocol: fieldObj.proto,
-                    sourceIp: fieldObj.srcIp,
-                    destinationIp: fieldObj.dstIp,
-                    data: packet_data,
-                })
-                let ethernetPacket = encodeEthernetFrame({
-                    data: ip_packet,
-                    type: ipv4_regex.test(fieldObj.dstIp) ? '0800' : '86dd'
-                }) 
-                if (proto === 'TCP') {
-                    sequenceMap.set(hash, (sequenceMap.get(hash) ?? 1) + Buffer.from(line).length)
-                }
-                return {
-                    timestamp: fieldObj.ts,
-                    buffer: ethernetPacket,
-                    type: fields[0]?.values[index]?.type
-                }
-            })
-            const generator = configure({ Buffer: Buffer, snapshotLength: 102400, linkLayerType: 1 })
-            const pcapFile = generator(packets2)
-            const blob = new Blob([pcapFile], { type: 'application/vnd.tcpdump.pcap' });
-            // // Create element with <a> tag
-            const link = document.createElement("a");
-
-            // Add file content in the object URL
-            link.href = URL.createObjectURL(blob);
-
-            // Add file name
-            const date = new Date();
-            link.download = `${convertDateToFileName(date)}.pcap`;
-            link.click();
-
+            pcapExporter(data);
         }
         document.addEventListener('export-flow-as-pcap', handler);
         return () => {
@@ -397,57 +235,8 @@ export const SimplePanel: React.FC<Props> = ({ options, data, width, height }) =
     const [filters, setFilters] = useState<Filters>({ ip: {}, port: {}, ipPort: {}, method: {}, type: {}, callid: {} });
     // Set flow data and sort
     useEffect(() => {
-        console.log(filters)
-        const [serie]: any = (data as any)?.series || [];
-        const fields = serie?.fields || [];
-        if (fields) {
-            const [firsField]: any = fields;
-            const sortData = formattingDataAndSortIt(fields, options.sortoption);
-            const outData = firsField?.values;
-            const map = new Map();
-            if (outData) {
-                setFlowData({
-                    actors: [], data: sortData.map((item: any) => {
-                        const message: string = item.Line || '';
-                        const labelItem: any = item.labels || {};
-                        const getOptionValue = (optionArr: string[] | string) => {
-                            if (optionArr instanceof Array) {
-                                return optionArr.map((option: string) => labelFormatter(labelItem[option], option)).filter((a: any) => !!a).join(':');
-                            }
-                            return labelItem[optionArr] || '';
-                        };
-                        const itemHash = hash(JSON.stringify(item))
-                        map.set(itemHash, item);
-
-                        const isSrcIPDisabled = !(filters?.['ip']?.[labelItem.src_ip] ?? true)
-                        const isDstIPDisabled = !(filters?.['ip']?.[labelItem.dst_ip] ?? true)
-                        const isSrcPortDisabled = !(filters?.['port']?.[labelItem.src_port] ?? true)
-                        const isDstPortDisabled = !(filters?.['port']?.[labelItem.dst_port] ?? true)
-                        const isSrcIpPortDisabled = !(filters?.['ipPort']?.[labelItem.src_ip + ':' + labelItem.src_port] ?? true)
-                        const isDstIpPortDisabled = !(filters?.['ipPort']?.[labelItem.dst_ip + ':' + labelItem.dst_port] ?? true)
-                        const isMethodDisabled = !(filters?.['method']?.[labelItem.response] ?? true)
-                        const isTypeDisabled = !(filters?.['type']?.[labelItem.type] ?? true)
-                        const isCallidDisabled = !(filters?.['callid']?.[labelItem.callid] ?? true)
-                        const hidden = isSrcIPDisabled || isDstIPDisabled || isSrcPortDisabled || isDstPortDisabled || isSrcIpPortDisabled || isDstIpPortDisabled || isMethodDisabled || isTypeDisabled || isCallidDisabled
-
-                        return {
-                            messageID: getOptionValue(options.colorGenerator) || 'Title',
-                            details: getOptionValue(options.details) || '',
-                            line: options.showbody && message || '',
-                            source: getOptionValue(options.source) || '...',
-                            destination: getOptionValue(options.destination) || '...',
-                            title: getOptionValue(options.title) || '',
-                            aboveArrow: getOptionValue(options.aboveArrow) || '',
-                            belowArrow: getOptionValue(options.belowArrow) || '',
-                            sourceLabel: getOptionValue(options.sourceLabel) || '',
-                            destinationLabel: getOptionValue(options.destinationLabel) || '',
-                            hidden,
-                            hash: itemHash
-                        }
-                    }).filter((item: any) => !item.hidden)
-                })
-                setModalDataFields(map);
-            }
+        if (data && options) {
+            filterFlowItems(data, options, setFlowData, setModalDataFields, filters);
         }
     }, [data, options, filters]);
     console.log(flowData)
